requirements.txt
streamlit
pandas
scikit-learn
numpy
IMDbPY

